backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "containermaster-backend"
version = "0.1.0"
description = "Python backend for ContainerMaster"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.34,<1.0",
  "docker>=7.1,<8.0",
  "psutil>=7.0,<8.0",
  "PyJWT>=2.10,<3.0",
  "bcrypt>=4.1,<5.0",
  "python-dotenv>=1.0,<2.0",
]

[project.optional-dependencies]
dev = []

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
